backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
duckduckgo-search
sse-starlette